[squash]: Fix incorrect use of the _PIPE() macro
_PIPE(pipe, a, b) does ((a) + (pipe)*((b)-(a)))

One needs to give the pipe A and pipe B register as argument. It works
for pipe C if the offset between the pipe A and pipe B register ((b) - (a)) is
also the offset between the pipe B and pipe C versions for the register,
which is usually the case.