drm/i915: Implement blocking read for pipe CRC files
seq_file is not quite the right interface this one. We have a circular
buffer with a new entry per vblank on one side and a process wanting to
dequeue the CRC with a read().

It's quite racy to wait for vblank in user land and then try to read a
pipe_crc file, sometimes the CRC interrupt hasn't been fired and we end
up with an EOF.

So, let's have the read on the pipe_crc file block until the interrupt
gives us a new entry. At that point we can wake the reading process.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>