drm/i915: Remove the Quanta special case
We now read out the FUSE_STRAP and SFUSE_STRAP registers, looking for
configurations with display fused off. Let's remove the Quanta special
case and rely on the programmed fuses to set num_pipes to 0.

This patch is untested and could use a good soul with such a device to
give it a go.

Cc: Ben Widawsky <benjamin.widawsky@intel.com>
Reviewed-by: Mika Kuoppala <mika.kuoppala@intel.com>
Reviewed-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>