drm/i915: Move num_plane to the intel_device_info structure
And rename it to num_sprites as this value doesn't count the primary plane.

This limit lives with num_pipes really, and now that dev_priv->info is writable
we can put it there instead.

While at it, introduce a intel_device_info_runtime_init() where we'll be able
to gather the device info fields at run-time.

v2: rename num_plane to num_sprites (Ville Syrjälä)

v3: Rebase on to of the INTEL_INFO() removal

Reviewed-by: Mika Kuoppala <mika.kuoppala@intel.com> (for v2)
Reviewed-by: Ville Syrjälä <ville.syrjala@linux.intel.com> (for v2)
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>